mejoro calendario y algunas clases cambio en el analizador alimentos creando clases nuevas para las imagenes y añado la funcionalidad para cambiar de páginas y ver más productos

+++ src/main/java/com/palmen/foodtracker/repositories/IUsuarioRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import com.palmen.foodtracker.models.Usuario;
 
 @Repository
-public interface IUsuarioRepository extends JpaRepository<Usuario, Integer> {
+public interface IUsuarioRepository extends JpaRepository<Usuario, Long> {
 
 	 Optional<Usuario> findByNombreUsuario(String nombreUsuario);
 }
